requirements.txt
beautifulsoup4
requests
pandas
sqlalchemy
python-dotenv
psycopg2